+++ src/main/resources/templates/qna/show.html
@@ -53,7 +53,7 @@
                             <p class="qna-comment-count"><strong>{{question.answersCount}}</strong>개의 의견</p>
                             <div class="qna-comment-slipp-articles">
                                 {{#question.answers}}
-                                <article class="article" id="answer-1406">
+                                <article class="article" id="{{id}}">
                                     <div class="article-header">
                                         <div class="article-header-thumb">
                                             <img src="https://graph.facebook.com/v2.3/1324855987/picture"
@@ -66,8 +66,8 @@
                                             </a>
                                         </div>
                                     </div>
-                                    <div class="article-doc comment-doc" style="white-space:pre;">
-                                        <p>{{contents}}</p>
+                                    <div class="article-doc comment-doc">
+                                        {{contents}}
                                     </div>
                                     <div class="article-util">
                                         <ul class="article-util-list">
@@ -79,28 +79,28 @@
                                                 <form class="form-delete" action="/questions/{{question.id}}/answers/{{id}}"
                                                       method="POST">
                                                     <input type="hidden" name="_method" value="DELETE">
-                                                    <button type="submit" class="delete-answer-button">삭제</button>
+                                                    <button type="submit" class="delete-answer-button" name="/questions/{{question.id}}/answers/{{id}}">삭제</button>
                                                 </form>
                                             </li>
                                         </ul>
                                     </div>
                                 </article>
                                 {{/question.answers}}
-                                <form class="submit-write" method="post" action="/questions/{{question.id}}/answers">
-                                    <div class="form-group" style="padding:14px;">
-                                        <textarea name="contents" class="form-control" placeholder="Update your status"></textarea>
-                                    </div>
-                                    <button class="btn btn-success pull-right" type="submit">답변하기</button>
-                                    <div class="clearfix"/>
-                                </form>
                             </div>
+                            <form class="submit-write" method="post" action="/questions/{{question.id}}/answers">
+                                <div class="form-group" style="padding:14px;">
+                                    <textarea name="contents" class="form-control" placeholder="Update your status"></textarea>
+                                </div>
+                                <button class="btn btn-success pull-right" type="submit">답변하기</button>
+                                <div class="clearfix"/>
+                            </form>
                         </div>
                     </div>
                 </div>
             </div>
         </div>
     </div>
-
+<!--
     <script type="text/template" id="answerTemplate">
         <article class="article">
             <div class="article-header">
@@ -130,7 +130,9 @@
             </div>
         </article>
     </script>
+    -->
 
     {{> base/footer}}
+    <script src="/js/qna.js"></script>
 </body>
 </html>
